requirements.txt
pypresence
dbus-python
requests
